requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiofiles==24.1.0
sqlalchemy==2.0.35
fastapi==0.115.2
uvicorn[standard]==0.30.6
